pseudo: Verify build completion
Execute finishing scripts (md5, copy, upload) only if the build was successful
ie. actually produced a .zip file

+++ build_device.sh
@@ -52,36 +52,41 @@ else
     ZIP=$(basename $(grep "Package OTA" "$LOG_DIR"/"$TARGET_PRODUCT"_"$DATE"_bot.log | cut -f3 -d ' '))
 fi
 
-mkdir ../upload
-OUTD=$(echo $(cd ../upload && pwd))
-rm $OUTD/$ZIP
-if [ -n "$3" ]; then
-    NZIP="$TARGET_PRODUCT"_jb-mr1_"$3".zip
-    cp "$ANDROID_PRODUCT_OUT"/$ZIP $OUTD/$NZIP
-else
-    cp "$ANDROID_PRODUCT_OUT"/$ZIP $OUTD/$ZIP
-fi
+# execute finishing scripts (md5, upload, etc) only if the build was successful (ie. actually produced a .zip)
+if [[ $ZIP == *.zip* ]]; then
+    # finish
+    echo "$2 build complete"
+    mkdir ../upload
+    OUTD=$(echo $(cd ../upload && pwd))
+    rm $OUTD/$ZIP
+    if [ -n "$3" ]; then
+        NZIP="$TARGET_PRODUCT"_jb-mr1_"$3".zip
+        cp "$ANDROID_PRODUCT_OUT"/$ZIP $OUTD/$NZIP
+    else
+        cp "$ANDROID_PRODUCT_OUT"/$ZIP $OUTD/$ZIP
+    fi
 
-# finish
-echo "$2 build complete"
-
-# md5sum list
-cd $OUTD
-md5sum $ZIP | cat >> md5sum
-
-# upload
-echo "checking on upload reference file"
-
-BUILDBOT=$BUILD_ROOT/vendor/$TARGET_VENDOR/bot/
-cd $BUILDBOT
-if test -x upload ; then
-    echo "Upload file exists, executing now"
-    cp upload $OUTD
+    # md5sum list
     cd $OUTD
-    # device and zip names are passed on for upload
-    ./upload $2 $ZIP && rm upload
+    md5sum $ZIP | cat >> md5sum
+
+    # upload
+    echo "checking on upload reference file"
+
+    BUILDBOT=$BUILD_ROOT/vendor/$TARGET_VENDOR/bot/
+    cd $BUILDBOT
+    if test -x upload ; then
+        echo "Upload file exists, executing now"
+        cp upload $OUTD
+        cd $OUTD
+        # device and zip names are passed on for upload
+        ./upload $2 $ZIP && rm upload
+    else
+        echo "No upload file found (or set to +x), build complete."
+    fi
+
 else
-    echo "No upload file found (or set to +x), build complete."
+    echo "$2 build failed, exiting"
 fi
 
 cd $BUILD_ROOT
